requirements.txt
az511-client>=0.1.1
python-dotenv>=1.0.0
pytest>=8.0
pytest-mock>=3.12
